test: move appearance fixtures to their own conftest
The root conftest is loaded by every test, so anything in it looks like
it belongs to all of them. The appearance fakes only ever served the
appearance tests, and keeping them at the top made the shared file
bigger without telling anyone who needs what.

Putting them in the appearance folder keeps them close to the tests
that use them and leaves the shared conftest for things that really are
shared.
